+++ CHANGES.rst
@@ -15,6 +15,7 @@ general
 - Reduce total program weight from 13.0 MB to 2.9 MB by using lightweight
   unicode database from package ``readable-glyph-names`` for detecting diacritics
 - Make "Print view" the default
+- Use HtmlWebpackPlugin and CopyWebpackPlugin for improved bundling of HTML resources
 
 lqdn-gafam-campaign
 -------------------